refactor: JDBC에 적합한 테스트로 수정
